adapted to tests

+++ src/Controller/GameController.php
@@ -3,13 +3,13 @@
 namespace App\Controller;
 
 use App\Entity\Game;
+use App\Entity\User;
 use Doctrine\DBAL\Types\TextType;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 
 use Symfony\Component\Validator\Constraints as Assert;
 class GameController extends AbstractController
@@ -56,66 +56,82 @@ class GameController extends AbstractController
     }
 
     #[Route('/game/{identifiant}', name: 'fetch_game', methods:['GET'])]
-    public function getGameInfo(EntityManagerInterface $entityManager, int $identifiant): JsonResponse
+    public function getGameInfo(EntityManagerInterface $entityManager, $identifiant): JsonResponse
     {
-        $party = $entityManager->getRepository(Game::class)->findOneBy(['id' => $identifiant]);
+        if(ctype_digit($identifiant)){
+            $party = $entityManager->getRepository(Game::class)->findOneBy(['id' => $identifiant]);
 
-        if($party !== null){
-            return $this->json(
-                $party,
-                headers: ['Content-Type' => 'application/json;charset=UTF-8']
-            );
+            if($party !== null){
+                return $this->json(
+                    $party,
+                    headers: ['Content-Type' => 'application/json;charset=UTF-8']
+                );
+            }else{
+                return new JsonResponse('Game not found', 404);
+            }
         }else{
             return new JsonResponse('Game not found', 404);
         }
     }
 
     #[Route('/games/{id}/add/{playerRightId}', name: 'add_user_right', methods:['PATCH'])]
-    public function inviteToGame(Request $request, EntityManagerInterface $entityManager, int $id, int $playerRightId): JsonResponse
+    public function inviteToGame(Request $request, EntityManagerInterface $entityManager, $id, $playerRightId): JsonResponse
     {
         $currentUserId = $request->headers->get('X-User-Id');
-        $playerLeft = $entityManager->getRepository(User::class)->find($currentUserId);
 
-        if($playerLeft === null){
-            return new JsonResponse('User not found', 404);
-        }
+        if(ctype_digit($id) && ctype_digit($playerRightId) && ctype_digit($currentUserId)){
+            $playerLeft = $entityManager->getRepository(User::class)->find($currentUserId);
 
-        $game = $entityManager->getRepository(Game::class)->find($id);
-
-        if($game === null){
-            return new JsonResponse('Game not found', 404);
-        }
-
-        if($game->getState() === 'ongoing' || $game->getState() === 'finished'){
-            return new JsonResponse('Game already started', 409);
-        }
-
-        $playerRight = $entityManager->getRepository(User::class)->find($playerRightId);
-
-        if($playerRight !== null){
-
-            if($playerLeft->getId() === $playerRight->getId()){
-                return new JsonResponse('You can\'t play against yourself', 409);
+            if($playerLeft === null){
+                return new JsonResponse('User not found', 404);
             }
-            
-            $game->setPlayerRight($playerRight);
-            $game->setState('ongoing');
 
-            $entityManager->flush();
+            $game = $entityManager->getRepository(Game::class)->find($id);
 
-            return $this->json(
-                $game,
-                headers: ['Content-Type' => 'application/json;charset=UTF-8']
-            );
+            if($game === null){
+                return new JsonResponse('Game not found', 404);
+            }
+
+            if($game->getState() === 'ongoing' || $game->getState() === 'finished'){
+                return new JsonResponse('Game already started', 409);
+            }
+
+            $playerRight = $entityManager->getRepository(User::class)->find($playerRightId);
+
+            if($playerRight !== null){
+
+                if($playerLeft->getId() === $playerRight->getId()){
+                    return new JsonResponse('You can\'t play against yourself', 409);
+                }
+                
+                $game->setPlayerRight($playerRight);
+                $game->setState('ongoing');
+
+                $entityManager->flush();
+
+                return $this->json(
+                    $game,
+                    headers: ['Content-Type' => 'application/json;charset=UTF-8']
+                );
+            }else{
+                return new JsonResponse('User not found', 404);
+            }
         }else{
-            return new JsonResponse('User not found', 404);
+            return new JsonResponse('Game not found', 404);
         }
     }
 
     #[Route('/game/{identifiant}', name: 'send_choice', methods:['PATCH'])]
     public function play(Request $request, EntityManagerInterface $entityManager, int $identifiant): JsonResponse
     {
+        if(ctype_digit($identifiant) === false){
+            return new JsonResponse('Game not found', 404);
+        }
         $currentUserId = $request->headers->get('X-User-Id');
+
+        if(ctype_digit($currentUserId) === false){
+            return new JsonResponse('User not found', 404);
+        }
 
         $currentUser = $entityManager->getRepository(User::class)->find($currentUserId);
 
@@ -166,34 +182,34 @@ class GameController extends AbstractController
             }
 
             if($userIsPlayerLeft){
-                $game->setPlayerLeftChoice($data['choice']);
+                $game->setPlayLeft($data['choice']);
                 $entityManager->flush();
 
-                if($game->getPlayerRightChoice() !== null){
+                if($game->getPlayRight() !== null){
                         
                     switch($data['choice']){
                         case 'rock':
-                            if($game->getPlayerRightChoice() === 'paper'){
+                            if($game->getPlayRight() === 'paper'){
                                 $game->setResult('winRight');
-                            }elseif($game->getPlayerRightChoice() === 'scissors'){
+                            }elseif($game->getPlayRight() === 'scissors'){
                                 $game->setResult('winLeft');
                             }else{
                                 $game->setResult('draw');
                             }
                             break;
                         case 'paper':
-                            if($game->getPlayerRightChoice() === 'scissors'){
+                            if($game->getPlayRight() === 'scissors'){
                                 $game->setResult('winRight');
-                            }elseif($game->getPlayerRightChoice() === 'rock'){
+                            }elseif($game->getPlayRight() === 'rock'){
                                 $game->setResult('winLeft');
                             }else{
                                 $game->setResult('draw');
                             }
                             break;
                         case 'scissors':
-                            if($game->getPlayerRightChoice() === 'rock'){
+                            if($game->getPlayRight() === 'rock'){
                                 $game->setResult('winRight');
-                            }elseif($game->getPlayerRightChoice() === 'paper'){
+                            }elseif($game->getPlayRight() === 'paper'){
                                 $game->setResult('winLeft');
                             }else{
                                 $game->setResult('draw');
@@ -211,7 +227,7 @@ class GameController extends AbstractController
                 }
 
             }elseif($userIsPlayerRight){            
-                $game->setPlayerRightChoice($data['choice']);
+                $game->setPlayRight($data['choice']);
 
                 $entityManager->flush();
 
@@ -227,31 +243,31 @@ class GameController extends AbstractController
 
 
 
-                if($game->getPlayerLeftChoice() !== null){
+                if($game->getPlayLeft() !== null){
 
                     switch($data['choice']){
                         case 'rock':
-                            if($game->getPlayerLeftChoice() === 'paper'){
+                            if($game->getPlayLeft() === 'paper'){
                                 $game->setResult('winLeft');
-                            }elseif($game->getPlayerLeftChoice() === 'scissors'){
+                            }elseif($game->getPlayLeft() === 'scissors'){
                                 $game->setResult('winRight');
                             }else{
                                 $game->setResult('draw');
                             }
                             break;
                         case 'paper':
-                            if($game->getPlayerLeftChoice() === 'scissors'){
+                            if($game->getPlayLeft() === 'scissors'){
                                 $game->setResult('winLeft');
-                            }elseif($game->getPlayerLeftChoice() === 'rock'){
+                            }elseif($game->getPlayLeft() === 'rock'){
                                 $game->setResult('winRight');
                             }else{
                                 $game->setResult('draw');
                             }
                             break;
                         case 'scissors':
-                            if($game->getPlayerLeftChoice() === 'rock'){
+                            if($game->getPlayLeft() === 'rock'){
                                 $game->setResult('winLeft');
-                            }elseif($game->getPlayerLeftChoice() === 'paper'){
+                            }elseif($game->getPlayLeft() === 'paper'){
                                 $game->setResult('winRight');
                             }else{
                                 $game->setResult('draw');
@@ -275,28 +291,36 @@ class GameController extends AbstractController
     #[Route('/game/{id}', name: 'annuler_game', methods:['DELETE'])]
     public function deleteGame(EntityManagerInterface $entityManager, Request $request, int $id): JsonResponse
     {
+        if(ctype_digit($id) === false){
+            return new JsonResponse('Game not found', 404);
+        }
+
         $currentUserId = $request->headers->get('X-User-Id');
-        $player = $entityManager->getRepository(User::class)->find($currentUserId);
 
-        if($player !== null){
-            $game = $entityManager->getRepository(Game::class)->findOneBy(['id' => $id, 'playerLeft' => $player]);
+        if(ctype_digit($currentUserId) === true){
+            $player = $entityManager->getRepository(User::class)->find($currentUserId);
 
-            if(empty($game)){
-                $game = $entityManager->getRepository(Game::class)->findOneBy(['id' => $id, 'playerRight' => $player]);
+            if($player !== null){
+                $game = $entityManager->getRepository(Game::class)->findOneBy(['id' => $id, 'playerLeft' => $player]);
+
+                if(empty($game)){
+                    $game = $entityManager->getRepository(Game::class)->findOneBy(['id' => $id, 'playerRight' => $player]);
+                }
+
+                if(empty($game)){
+                    return new JsonResponse('Game not found', 404);
+                }
+
+                $entityManager->remove($game);
+                $entityManager->flush();
+
+                return new JsonResponse(null, 204);
+
+            }else{
+                return new JsonResponse('User not found', 404);
             }
-
-            if(empty($game)){
-                return new JsonResponse('Game not found', 404);
-            }
-
-            $entityManager->remove($game);
-            $entityManager->flush();
-
-            return new JsonResponse(null, 204);
-
         }else{
             return new JsonResponse('User not found', 404);
         }
-
     }
 }
